vm: fix array_set_field_double()

+++ include/vm/object.h
@@ -181,9 +181,9 @@ array_get_field_float(struct vm_object *obj, int index)
 }
 
 static inline void
-array_set_field_double(struct vm_object *obj, int index, uint16_t value)
+array_set_field_double(struct vm_object *obj, int index, double value)
 {
-	*(uint16_t *) &obj->fields[index * get_vmtype_size(J_DOUBLE)] = value;
+	*(double *) &obj->fields[index * get_vmtype_size(J_DOUBLE)] = value;
 }
 
 static inline double
